Split spec common actions support helper
