+++ spec/support/common_actions/tags.rb
@@ -1,0 +1,6 @@
+module Tags
+  # spec/models/tag_cloud_spec.rb
+  def tag_names(tag_cloud)
+    tag_cloud.tags.map(&:name)
+  end
+end
